backend/cloud_functions/fetch_fitness_data/requirements.txt
google-cloud-storage
flask<3.0,>=1.0
flask-cors 
# dotenv # for environment variables for local testing